DTSSTCI-86: Create utility event for testing purposes that allows test case creation in any chosen state
