+++ src/main/java/uk/gov/hmcts/sptribs/ciccase/model/CaseData.java
@@ -56,6 +56,7 @@ import static java.lang.String.format;
 import static java.time.format.DateTimeFormatter.ofPattern;
 import static java.util.Locale.UK;
 import static uk.gov.hmcts.ccd.sdk.type.FieldType.Collection;
+import static uk.gov.hmcts.ccd.sdk.type.FieldType.FixedList;
 import static uk.gov.hmcts.ccd.sdk.type.FieldType.FixedRadioList;
 import static uk.gov.hmcts.ccd.sdk.type.FieldType.TextArea;
 
@@ -424,6 +425,14 @@ public class CaseData {
     @CCD(access = {DefaultAccess.class})
     @JsonUnwrapped
     private RetiredFields retiredFields;
+
+    @CCD(
+        label = "State to transition test case to",
+        typeOverride = FixedList,
+        typeParameterOverride = "State",
+        access = {DefaultAccess.class}
+    )
+    private State stateToTransitionApplicationTo;
 
     public String getFirstHearingDate() {
